Bug fixes and mechanics updates

+++ include/move_table.h
@@ -19,7 +19,8 @@ typedef struct MoveTable {
     u8 effectChance;
     u16 range;
     s8 priority;
-    u8 flags;
+    // phmode: widened from u8 to u16 to fit MOVE_FLAG_POWDER, the 9th move flag.
+    u16 flags;
     MoveContestData contest;
 } MoveTable;
